+++ gems/aws-sdk-core/lib/aws-sdk-core/plugins/retry_errors.rb
@@ -72,6 +72,7 @@ A delay randomiser function used by the default backoff function. Some predefine
           'UnrecognizedClientException', # json services
           'InvalidAccessKeyId',          # s3
           'AuthFailure',                 # ec2
+          'InvalidIdentityToken',        # sts
         ])
 
         THROTTLING_ERRORS = Set.new([
@@ -117,6 +118,7 @@ A delay randomiser function used by the default backoff function. Some predefine
 
         def networking?
           @error.is_a?(Seahorse::Client::NetworkingError) ||
+          @error.is_a?(Errors::NoSuchEndpointError) ||
           NETWORKING_ERRORS.include?(@name)
         end