+++ gems/aws-sdk-core/lib/aws-sdk-sts/client.rb
@@ -1794,7 +1794,7 @@ module Aws::STS
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-core'
-      context[:gem_version] = '3.61.1'
+      context[:gem_version] = '3.61.2'
       Seahorse::Client::Request.new(handlers, context)
     end